// src/app/models/notification/notificationType.ts
export const enum NotificationType {
  MESSAGE = 'MESSAGE',
  ARBITRATION = 'ARBITRATION',
  ARBITRATION_USER_APPROVED = 'ARBITRATION_USER_APPROVED',
  ADMIN_REQUEST = 'ADMIN_REQUEST',
  TASK_CHANGED = 'TASK_CHANGED',
  TASK_CREATED = 'TASK_CREATED',
  COMMENT = 'COMMENT',
  ACTIVATE = 'ACTIVATE',
  BLOCKING = 'BLOCKING',
  DELETE_TASK = 'DELETE_TASK',
  BONUS = 'BONUS',
  DELETE_VACANCY='DELETE_VACANCY',
  DELETE_RESUME='DELETE_RESUME',
  VACANCY_CREATED="VACANCY_CREATED",
}